Animal Crossing Town Tickets v.1.2.6 by BrianMp16
Changelog:

Put Pinwheels and Fans in both the Seasonal and Other groups for the everything item filters.
Applied dual filters for Pinwheels, Fans, and Seasonal Nook Items to the Catalog as well as the Shop.
Losslessly reduced Resources folder size by ~20% using optipng tool.

This program is now under 100MB and should be downloadable through Github.